requirements.txt
tensorflow-gpu>=2.3.1
numpy
scipy